Remove previous incarnation
- Moved image processing functionality to cptv/image.py
- Removed parsingParsingThermaData.py (no longer needed)
